Champion super abstract class
